requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
anthropic>=0.52.0
httpx>=0.28.0
